% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellRangerLikeEmptyDrops.R
\name{cellRangerLikeEmptyDrops}
\alias{cellRangerLikeEmptyDrops}
\alias{.test_empty_drops_cr}
\alias{cellRangerLikeEmptyDrops,ANY-method}
\alias{cellRangerLikeEmptyDrops,SummarizedExperiment-method}
\title{An approximate implementation of the \code{--soloCellFilter  EmptyDrops_CR} filtering approach to identify empty droplets.}
\usage{
.test_empty_drops_cr(
  m,
  lower = 100,
  upper = NULL,
  niters = 10000,
  ignore = NULL,
  oversee = NULL,
  alpha = NULL,
  round = TRUE,
  by.rank.lower = NULL,
  by.rank.upper = NULL,
  BPPARAM = SerialParam()
)

cellRangerLikeEmptyDrops(m, ...)

\S4method{cellRangerLikeEmptyDrops}{ANY}(
  m,
  expected = 3000,
  max.percentile = 0.99,
  max.min.ratio = 10,
  umi.min = 500,
  umi.min.frac.median = 0.01,
  can.max.n = 20000,
  by.rank.upper = 45000,
  by.rank.lower = 90000,
  fdr.threshold = 0.01,
  barcode.args = list(),
  round = TRUE,
  ...,
  BPPARAM = SerialParam()
)

\S4method{cellRangerLikeEmptyDrops}{SummarizedExperiment}(m, ..., assay.type = "counts")
}
\arguments{
\item{m}{A numeric matrix-like object containing counts, where columns represent barcoded droplets and rows represent features.
The matrix should only contain barcodes for an individual sample, prior to any filtering for barcodes.}

\item{round}{Logical scalar indicating whether to check for non-integer values in \code{m} and, if present, round them for ambient profile estimation (see \code{?\link{ambientProfileEmpty}}) and the multinomial simulations.}

\item{by.rank.lower}{An integer specifying the highest UMI count ranking of the ambient pool, barcodes with UMI count ranking above
this number will not be included in the ambient pool. It is same as \code{indMax} in STARsolo.}

\item{by.rank.upper}{An integer specifying the lowest UMI count ranking of the ambient pool, barcodes with UMI count ranking below
this number will not be included in the ambient pool. It is same as \code{indMin} in STARsolo.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating whether parallelization should be used.}

\item{...}{For the generic, further arguments to pass to to individual methods.}

\item{expected}{A numeric scalar specifying the expected number of barcodes in this sample It is same as \code{nExpectedCells} in STARsolo.}

\item{max.percentile}{A numeric scalar specifying the percentile used in simple filtering, barcodes selected by simple filtering 
will be regarded as real barcodes regardless of the \code{emptyDrops} result. It is same as \code{maxPercentile} in STARsolo.}

\item{max.min.ratio}{A numeric scalar specifying the maximum ratio of maximum UMI count and minimum UMI count used in simple filtering, 
maximum UMI count used in simple filtering is determined first by \code{expected*(1-max.percentile)}, minimum UMI count used in
 simple filtering is then determined by this ratio. It is same as \code{maxMinRatio} in STARsolo.}

\item{umi.min}{A numeric scalar specifying the minimum UMI count above which a sample will be included in ambient profiles. It is same as \code{umiMin} in STARsolo.}

\item{umi.min.frac.median}{A numeric scalar between 0 and 1 specifying that only the barcodes whose UMI count are above this number 
fraction of the median UMI count of the top \code{expected} barcodes will be included in the ambient profile. 
It is same as \code{umiMinFracMedian} in STARsolo.}

\item{barcode.args}{Further arguments to pass to \code{\link{barcodeRanks}}.}

\item{cand.max.n}{An integer specifying the maximum number of ambient barcodes that are possible to be regarded as real barcodes. It is same as \code{canMaxN} in STARsolo.}

\item{fdr.thresh}{A numeric scalar specifying the FDR threshold to filter barcodes. barcodes whose FDR returned by emptyDrops
is above this threshold will not be regarded as real barcodes. It is same as \code{FDR} in STARsolo.}
}
\value{
A DataFrame like \code{\link{emptyDrops}}, with an additional binary \code{is.cell} field demonstrating whether
barcodes are estimated as real barcodes.
}
\description{
An approximate implementation of the \code{--soloCellFilter  EmptyDrops_CR} filtering approach, 
which, itself, was reverse-engineered from the behavior of CellRanger 3+.
}
\details{
This function is an approximate implementation of the  \code{--soloCellFilter  EmptyDrops_CR} filtering approach of STARsolo 2.7.9a
, which, itself, was reverse engineered from the behavior of  CellRanger 3+. 
All parameters are defaulty set as the default value used in starSolo 2.7.9a.
In the most cases, users just need to specify the raw and unfiltered count matrix, \code{m}.
See \code{?\link{emptyDrops}} for an alternative approach for cell calling.
}
\examples{
# Mocking up some data:
set.seed(0)
my.counts <- DropletUtils:::simCounts()

# Identify likely cell-containing droplets.
e.out <- cellRangerLikeEmptyDrops(my.counts)
e.out

# Get matrix of estimated barcodes.
cell.counts <- my.counts[, e.out$is.cell]

}
\references{
Kaminow et al. (2021).
STARsolo: accurate, fast and versatile mapping/quantification of single-cell and single-nucleus RNA-seq data
\url{https://www.biorxiv.org/content/10.1101/2021.05.05.442755v1}
}
\seealso{
\code{\link{emptyDrops}}, for another method for calling barcodes.
}
\author{
Dongze He, Rob Patro
}
