% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashedDrops.R
\name{hashedDrops}
\alias{hashedDrops}
\title{Demultiplex cell hashing data}
\usage{
hashedDrops(x, prop = NULL, pseudo.count = 3)
}
\arguments{
\item{x}{A numeric/integer matrix-like object containing UMI counts.
Rows correspond to HTO and columns correspond to cell barcodes.}

\item{ambient}{A numeric vector of length equal to \code{nrow(x)},
specifying the ambient (i.e., background) proportions of each HTO.}

\item{min.lfc}{Numeric scalar specifying the minimum log-fold change above the ambient noise when testing for significant HTOs.}

\item{lower}{Numeric scalar specifying the total count threshold for using barcodes to compute the ambient proportions.}

\item{ignore}{Numeric scalar specifying the total count threshold for ignoring barcodes.}
}
\value{
A \linkS4class{DataFrame} with one row per column of \code{x}, containing the following fields:
\itemize{
\item \code{Total}, integer specifying the total count across all HTOs for this barcode.
\item \code{Best}, integer specifying the best HTO with the lowest p-value.
\item \code{PValue}, numeric containing the p-value for the best HTO.
\item \code{FDR}, a Benjamini-Hochberg-adjusted p-value for the best HTO.
\item \code{Second}, integer specifying the HTO with the second-lowest p-value.
\item \code{PValue2}, numeric containing the p-value for the second-best HTO.
\item \code{FDR2}, a Benjamini-Hochberg-adjusted p-value for the second-best HTO.
\item \code{LogFC}, numeric containing the log-fold change between the counts of the best and second-best HTO.
}
If \code{ignore} is specified, all barcodes with total counts at or below \code{ignore} will have \code{NA} statistics.
}
\description{
Demultiplex cell barcodes into their samples of origin based on the most significant hash tag oligo (HTO).
Also identify potential doublets based on the presence of multiple significant HTOs.
}
\details{
Note that this function is still experimental and subject to change.


The idea behind cell hashing is that cells from the same sample are stained with reagent conjugated with a single HTO.
Cells are then mixed across multiple samples and subjected to droplet-based single-cell sequencing.
Cell barcode libraries can then be demultiplexed into individual samples based on whether their unique HTO is detected.

In this function, we test for differences between each cell barcode's HTO profile and that of the ambient solution.
The top HTO for each barcode is identified as that which has the lowest p-value from a binomial test,
using the total count for the barcode as the size and the ambient proportions as the probability.
This strategy accounts for potential imbalances in the HTO distribution in the ambient solution
that would confound a simple approach of taking the most abundant HTO.
We apply the Bonferroni correction across HTOs to obtain the reported \code{PValue}.

Once the top HTO is identified for a cell barcode, we remove it from the total count and the ambient proportions.
We then repeat the aforementioned test procedure to identify the second-best HTO,
the p-value for which can be used as evidence for the presence of a doublet.
Note that the \code{PValue2} field is set to the maximum of the p-values of the top and second-best HTOs,
as it would not be possible to detect a doublet for a barcode that is not even detected as a cell.

Currently, this function does not support combinatorial indexing schemes, 
i.e., where samples are represented by a unique \emph{combination} of HTOs.
}
\section{Estimating the ambient proportions}{

By default, the ambient proportions are roughly estimated by taking the average count across all barcodes in \code{x}.
This is probably fine in most cases as long as the rates of lysis/cell damage are consisten across samples
(especially if they have been pooled together for multiplexed processing).

A more appropriate estimate may be obtained by computing averages across barcodes with low total counts.
These barcodes are unlikely to contain cells and thus their profiles are more representative of the ambient solution.
(A similar approach is used in \code{\link{emptyDrops}} to obtain an estimate of the ambient profile.)
This can be achieved by setting \code{lower} to only use low-count barcodes for this calculation.
The definition of \dQuote{low} depends on the sequencing depth of the HTOs and so it is difficult to give a definitive guide,
but typical datasets might use \code{lower=1000}.

Alternatively, users can supply their own estimates of the ambient proportion to \code{ambient}.
Note that any vector will automatically be scaled to sum to unity.
}

\examples{
# Mocking up an example dataset with 10 HTOs and 10\% doublets.
ncells <- 1000
nhto <- 10
y <- matrix(rpois(ncells*nhto, 50), nrow=nhto)
true.sample <- sample(nhto, ncells, replace=TRUE)
y[cbind(true.sample, seq_len(ncells))] <- 1000

ndoub <- ncells/10
next.sample <- (true.sample[1:ndoub]  + 1) \%\% nrow(y)
next.sample[next.sample==0] <- nrow(y)
y[cbind(next.sample, seq_len(ndoub))] <- 500

# Computing statistics; in this case we expect empty droplets
# to have total counts below 1000.
stats <- hashedDrops(y)

}
\author{
Aaron Lun
}
