% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10xPeaks.R
\name{read10xPeaks}
\alias{read10xPeaks}
\title{Load ATAC-seq data from a 10X Genomics experiment}
\usage{
read10xPeaks(
  sample,
  type = c("auto", "sparse", "HDF5", "prefix"),
  compressed = NULL
)
}
\arguments{
\item{sample}{A string containing the path to a single directory corresponding to the output of a CellRanger run.
Alternatively, each string may contain a path to a HDF5 file in the sparse matrix format peakrated by 10X.}

\item{type}{String specifying the type of 10X format to read data from.}
}
\value{
A \linkS4class{SingleCellExperiment} object containing count data for each peak/motif (row) and cell (column).
\itemize{
\item Row metadata will either be a \code{\link{GRanges}} object if the features are peaks,
or a \linkS4class{DataFrame} of transcription factor motif identifiers and names.
\item Column metadata will contain the \code{"Barcode"} field, the cell barcode sequence for each cell library. 
\item Rows are named with the peak identifier, while columns are named with the cell barcode.
\item The assays will contain a single \code{"counts"} matrix, containing UMI counts for each peak/motif in each cell.
Note that the matrix representation will depend on the format of the \code{sample}, see Details.
}
}
\description{
Creates a \linkS4class{SingleCellExperiment} from the CellRanger output directories for 10X Genomics ATAC-seq data.
}
\details{
If \code{type="auto"}, the format of the input file is automatically detected for each \code{samples} based on whether it ends with \code{".h5"}.
If so, \code{type} is set to \code{"HDF5"}; otherwise it is set to \code{"sparse"}.
\itemize{
\item If \code{type="sparse"}, count data are loaded as a \linkS4class{dgCMatrix} object.
This is a conventional column-sparse compressed matrix format produced by the CellRanger pipeline,
consisting of a (possibly Gzipped) MatrixMarket text file (\code{"matrix.mtx"})
with additional tab-delimited files for barcodes (\code{"barcodes.tsv"})
and peak/motif annotation (\code{"peaks.bed"} or \code{"motifs.tsv"}).
\item If \code{type="prefix"}, count data are also loaded as a \linkS4class{dgCMatrix} object.
This assumes the same three-file structure for each sample as described for \code{type="sparse"},
but each sample is defined here by a prefix in the file names rather than by being a separate directory.
For example, if the \code{samples} entry is \code{"xyx_"},
the files are expected to be \code{"xyz_matrix.mtx"}, \code{"xyz_barcodes.tsv"}, etc.
\item If \code{type="HDF5"}, count data are assumed to follow the 10X sparse HDF5 format for large data sets.
It is loaded as a \linkS4class{TENxMatrix} object, which is a stub object that refers back to the file in \code{samples}.
}

When \code{type="sparse"} or \code{"prefix"} and \code{compressed=NULL},
the function will automatically search for both the unzipped and Gzipped versions of the files.
This assumes that the compressed files have an additional \code{".gz"} suffix.
We can restrict to only compressed or uncompressed files by setting \code{compressed=TRUE} or \code{FALSE}, respectively.
}
\examples{
# Mocking up some 10X genomics output.
example(write10xCounts)

# Reading it in.
sce10x <- read10xCounts(tmpdir)

# Column names are dropped with multiple 'samples'.
sce10x2 <- read10xCounts(c(tmpdir, tmpdir))

}
\seealso{
\code{\link{read10xCounts}}, to create the equivalent object for scRNA-seq data.
}
\author{
Aaron Lun
}
