\name{read10xResults}
\alias{read10xResults}

\title{Load in data from 10x experiment}
\description{
Creates a SingleCellExperiment from the CellRanger output files for 10X Genomics data.
}

\usage{
read10xResults(samples)
}

\arguments{
\item{samples}{A character vector specifying one or more directory names, each corresponding to a 10X sample.
Each directory should contain the \code{matrix.mtx}, \code{genes.tsv} and \code{barcodes.tsv} files generated by CellRanger.
}
}

\value{
A SingleCellExperiment object with counts data stored as a sparse matrix. 
Matrices are combined by column if multiple \code{samples} were specified.
Rows are named with the gene identifier (e.g., ENSEMBL).
Columns are named with the cell barcode if \code{sample} contains one element; otherwise the columns are unnamed to avoid problems with non-unique barcodes.
}

\details{
This function was developed from the \code{Read10X} function from the \pkg{Seurat} package.
}

\author{
Davis McCarthy, with modifications from Aaron Lun
}

\examples{
# Mocking up some 10X genomics output.
tmpdir <- tempfile()
dir.create(tmpdir)

library(Matrix)
my.counts <- matrix(rpois(1000, lambda=5), ncol=10, nrow=100)
my.counts <- as(my.counts, "dgCMatrix")
writeMM(my.counts, file=file.path(tmpdir, "matrix.mtx"))

ngenes <- nrow(my.counts)
gene.ids <- paste0("GENE", seq_len(ngenes))
gene.symb <- paste0(sample(LETTERS, replace=TRUE, ngenes),
    sample(LETTERS, replace=TRUE, ngenes),
    sample(LETTERS, replace=TRUE, ngenes), "-",
    sample(9, replace=TRUE, ngenes))
write.table(data.frame(gene.ids, gene.symb), file=file.path(tmpdir, "genes.tsv"),
    quote=FALSE, col.names=FALSE, row.names=FALSE)   

cell.ids <- paste0("BARCODE-", seq_len(ncol(my.counts)))
write(cell.ids, file=file.path(tmpdir, "barcodes.tsv"))

# Reading it in.
sce10x <- read10xResults(tmpdir)

# Column names are dropped with multiple 'samples'.
sce10x2 <- read10xResults(c(tmpdir, tmpdir))
}
