% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeRanks.R
\name{barcodeRanks}
\alias{barcodeRanks}
\alias{barcodeRanks,ANY-method}
\alias{barcodeRanks,SummarizedExperiment-method}
\title{Calculate barcode ranks}
\usage{
barcodeRanks(m, ...)

\S4method{barcodeRanks}{ANY}(
  m,
  lower = 100,
  exclude.from = 50,
  fit.bounds = NULL,
  df = 20,
  ...,
  BPPARAM = SerialParam()
)

\S4method{barcodeRanks}{SummarizedExperiment}(m, ..., assay.type = "counts")
}
\arguments{
\item{m}{A numeric matrix-like object containing UMI counts, where columns represent barcoded droplets and rows represent genes.
Alternatively, a \linkS4class{SummarizedExperiment} containing such a matrix.}

\item{...}{For the generic, further arguments to pass to individual methods.

For the SummarizedExperiment method, further arguments to pass to the ANY method.}

\item{lower}{A numeric scalar specifying the lower bound on the total UMI count, 
at or below which all barcodes are assumed to correspond to empty droplets and excluded from knee/inflection point identification.}

\item{exclude.from}{An integer scalar specifying the number of highest ranking barcodes to exclude from knee/inflection point identification.}

\item{fit.bounds, df}{Deprecated and ignored.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying how parallelization should be performed.}

\item{assay.type}{Integer or string specifying the assay containing the count matrix.}
}
\value{
A \linkS4class{DataFrame} where each row corresponds to a column of \code{m}, and containing the following fields:
\describe{
\item{\code{rank}:}{Numeric, the rank of each barcode (averaged across ties).}
\item{\code{total}:}{Numeric, the total counts for each barcode.}
}

The metadata contains \code{knee}, a numeric scalar containing the total count at the knee point;
and \code{inflection}, a numeric scalar containing the total count at the inflection point.
}
\description{
Compute barcode rank statistics and identify the knee and inflection points on the total count curve.
}
\details{
Analyses of droplet-based scRNA-seq data often show a plot of the log-total count against the log-rank of each barcode
where the highest ranks have the largest totals.
This is equivalent to a transposed empirical cumulative density plot with log-transformed axes, 
which focuses on the barcodes with the largest counts.
To help create this plot, the \code{barcodeRanks} function will compute these ranks for all barcodes in \code{m}.
Barcodes with the same total count receive the same average rank to avoid problems with discrete runs of the same total.

The function will also identify the inflection and knee points on the curve for downstream use.
Both of these points correspond to a sharp transition between two components of the total count distribution, 
presumably reflecting the difference between empty droplets with little RNA and cell-containing droplets with much more RNA.
\itemize{
\item The inflection point is defined as the point on the log-rank/log-total curve where the first derivative is minimized.
If multiple inflection points are present, we choose the point that immediately follows the knee point.
\item To find the knee point, we draw a diagonal line that passes through the inflection point in the log-rank/log-total curve.
The knee point is defined as the location on the curve that is above and most distant from this line.
}
Only points with total counts above \code{lower} will be considered for knee/inflection point identification.
Similarly, the first \code{exclude.from} points will be ignored to avoid instability at the start of the curve.
}
\examples{
# Mocking up some data: 
set.seed(2000)
my.counts <- DropletUtils:::simCounts()

# Computing barcode rank statistics:
br.out <- barcodeRanks(my.counts)
names(br.out)

# Making a plot.
plot(br.out$rank, br.out$total, log="xy", xlab="Rank", ylab="Total")
o <- order(br.out$rank)
abline(h=metadata(br.out)$knee, col="dodgerblue", lty=2)
abline(h=metadata(br.out)$inflection, col="forestgreen", lty=2)
legend("bottomleft", lty=2, col=c("dodgerblue", "forestgreen"), 
    legend=c("knee", "inflection"))

}
\seealso{
\code{\link{emptyDrops}}, where this function is used.
}
\author{
Aaron Lun
}
