\name{read10xCounts}
\alias{read10xCounts}

\title{Load in data from 10x experiment}

\description{
Creates a SingleCellExperiment from the CellRanger output directories for 10X Genomics data.
}

\usage{
read10xCounts(samples, col.names=FALSE, ...)
}

\arguments{
\item{samples}{A character vector containing one or more directory names, each corresponding to a 10X sample.
Each directory should contain the \code{"matrix.mtx"}, \code{"genes.tsv"} and \code{"barcodes.tsv"} files generated by CellRanger.
}
\item{col.names}{A logical scalar indicating whether the columns of the output object should be named with the cell barcodes.}
\item{...}{Further arguments to pass to \code{\link{read10xMatrix}}.}
}

\value{
A SingleCellExperiment object with counts data stored as a sparse matrix. 
Matrices are combined by column if multiple \code{samples} were specified.
Rows are named with the gene identifier (usually ENSEMBL).
}

\details{
This function was originally developed from the \code{Read10X} function from the \pkg{Seurat} package.
It was then taken from the \code{read10xResults} implementation in the \pkg{scater} package.

If \code{col.names=TRUE} and \code{length(sample)==1}, each column is named by the cell barcode.
For multiple samples, the columns are unnamed to avoid problems with non-unique barcodes across samples.

Note that user-level manipulation of sparse matrices requires loading of the \pkg{Matrix} package.
Otherwise, calculation of \code{rowSums}, \code{colSums}, etc. will result in errors.
}

\author{
Davis McCarthy, with modifications from Aaron Lun
}

\seealso{
\code{\link{write10xCounts}}
}

\examples{
# Mocking up some 10X genomics output.
example(write10xCounts)

# Reading it in.
sce10x <- read10xCounts(tmpdir)

# Column names are dropped with multiple 'samples'.
sce10x2 <- read10xCounts(c(tmpdir, tmpdir))
}

\references{
Zheng GX, Terry JM, Belgrader P, and others (2017).
Massively parallel digital transcriptional profiling of single cells. 
\emph{Nat Commun} 8:14049.

10X Genomics (2017).
Gene-Barcode Matrices.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/matrices}
}
